### 2021-03-28 #3123 @Kevin-Mattheus-Moerman
@pibion that is fine. We now tag submissions as "waitlisted" if we feel the editors need time before they can commit fully to the submission. I can do two things, so let me know if you prefer A or B, option A) I could assign you here as editor and add the waitlisted label which you can remove once you have time to take this up, or B) I could not assign you yet and keep this waitlisted for when you are ready to be assigned (that way it doesn't show up on your list yet and may prevent that feeling of pressure).
