### 2021-04-14 #3171 @whedon
[REVIEW]: Surfaxe: Systematic surface calculations
**Submitting author:** @brlec (<a href="http://orcid.org/0000-0003-1485-1888">Katarina Brlec</a>)
**Repository:** <a href="https://github.com/SMTG-UCL/surfaxe" target ="_blank">https://github.com/SMTG-UCL/surfaxe</a>
**Version:** v1.2
