+++ surfaxe/generation.py
@@ -1,6 +1,7 @@
 # pymatgen
 from pymatgen.core.surface import SlabGenerator, generate_all_slabs
 from pymatgen.core import Structure
+from pymatgen.core.structure import SiteCollection
 from pymatgen.symmetry.analyzer import SpacegroupAnalyzer
 
 # misc
@@ -9,13 +10,15 @@ import os
 import itertools
 import functools
 import multiprocessing
+import math
+import numpy as np 
 
 # surfaxe
 from surfaxe.io import slabs_to_file, _custom_formatwarning
 
 def get_slabs_single_hkl(structure, hkl, thicknesses, vacuums, make_fols=False, 
 make_input_files=False, max_size=500, center_slab=True, ox_states=None, 
-save_slabs=True, is_symmetric=True, fmt='poscar', name='POSCAR',
+save_slabs=True, is_symmetric=True, layers_to_relax = None, fmt='poscar', name='POSCAR', 
 config_dict='PBEsol_config.json', user_incar_settings=None, 
 user_kpoints_settings=None, user_potcar_settings=None, **kwargs):
     """
@@ -193,9 +196,9 @@ user_kpoints_settings=None, user_potcar_settings=None, **kwargs):
                     if slab.is_symmetric() and not slab.is_polar():
                         provisional.append({
                             'hkl': ''.join(map(str, slab.miller_index)),
-                            'slab_t': thickness,
-                            'vac_t': vacuum,
-                            's_index': i,
+                            'slab_thickness': thickness,
+                            'vac_thickness': vacuum,
+                            'slab_index': i,
                             'slab': slab})
                 
                 # Get all the zero-dipole slabs wihtout inversion symmetry
@@ -203,15 +206,20 @@ user_kpoints_settings=None, user_potcar_settings=None, **kwargs):
                     if not slab.is_polar():
                         provisional.append({
                             'hkl': ''.join(map(str, slab.miller_index)),
-                            'slab_t': thickness,
-                            'vac_t': vacuum,
-                            's_index': i,
+                            'slab_thickness': thickness,
+                            'vac_thickness': vacuum,
+                            'slab_index': i,
                             'slab': slab})
                   
     # Iterate though provisional slabs to extract the unique slabs
     unique_list_of_dicts, repeat, large = _filter_slabs(provisional, max_size)
 
-    # Warnings for large, repeated and no slabs
+    if layers_to_relax is not None: 
+        unique_list_of_dicts, small = _get_selective_dynamics_single_hkl(
+            struc, unique_list_of_dicts, layers_to_relax
+        )
+
+    # Warnings for too large, too small, repeated and no slabs
     if repeat:
         warnings.formatwarning = _custom_formatwarning
         warnings.warn('Not all combinations of hkl or slab/vac thicknesses '
@@ -222,6 +230,12 @@ user_kpoints_settings=None, user_potcar_settings=None, **kwargs):
         warnings.formatwarning = _custom_formatwarning
         warnings.warn('Some generated slabs exceed the max size specified.'
         ' Slabs that exceed the max size are: ' + ', '.join(map(str, large)))
+    
+    if small: 
+        warnings.formatwarning = _custom_formatwarning
+        warnings.warn('Some slabs were too thin to fix the centre of the slab.'
+        ' Slabs with no selective dynamics applied are: ' + 
+        ', '.join(map(str, small)))
     
     if len(unique_list_of_dicts) == 0: 
         warnings.formatwarning = _custom_formatwarning
@@ -412,9 +426,9 @@ user_kpoints_settings=None, **kwargs):
                     if slab.is_symmetric() and not slab.is_polar():
                         provisional.append({
                             'hkl': ''.join(map(str, slab.miller_index)),
[…]
