### 2021-05-13 165c887
refactored generation of slabs, added some tests
