+++ tests/tests_generation/test_generation.py
@@ -5,7 +5,7 @@ import shutil
 from pathlib import Path
 from pymatgen.core import Structure
 from pymatgen.core.surface import Slab
-from surfaxe.generation import get_slabs_max_index, get_slabs_single_hkl
+from surfaxe.generation import get_slabs_max_index, get_slabs_single_hkl, generate_slabs
 
 ytos = str(Path(__file__).parents[2].joinpath('example_data/generation/CONTCAR_conventional'))
 cdte  = str(Path(__file__).parents[2].joinpath('example_data/generation/CdTe.vasp'))
@@ -102,6 +102,77 @@ class GetOneTestCase(unittest.TestCase):
     
     def test_save_to_file(self): 
         ytos_slabs = get_slabs_single_hkl(structure=self.ytos, 
+        hkl=(0,0,1), thicknesses=[10], vacuums=[10, 20])
+        
+        self.assertIsNone(ytos_slabs)
+        
+        # Check the files created 
+        self.assertEqual(len(os.listdir('Y4Ti4S4O10')), 8)
+        self.assertIn('POSCAR_001_10_10_15.vasp', os.listdir('Y4Ti4S4O10'))
+
+        # Clean up - get rid of directory created 
+        shutil.rmtree('Y4Ti4S4O10')
+
+class GenerateSlabsTestCase(unittest.TestCase): 
+
+    def setUp(self): 
+        self.ytos = ytos 
+        self.cdte = cdte
+    
+    def test_get_single_hkl(self): 
+        ytos_slab = generate_slabs(structure=self.ytos, hkl=(0,0,1), 
+        thicknesses=[10,20], vacuums=[10,20], save_slabs=False, max_size=20)
+
+        self.assertEqual(len(ytos_slab), 1)
+        self.assertEqual(ytos_slab[0]['s_index'], 4)
+        self.assertWarnsRegex(UserWarning, ('Some generated slabs exceed the '
+        'max size specified. Slabs that exceed the max size are: 001_10_10_4'))
+        self.assertWarnsRegex(UserWarning, ('Not all combinations of hkl or '
+        'slab/vac thicknesses were generated because of repeat structures. '
+        'The repeat slabs are: 001_20_10_4, 001_10_20_4, 001_20_20_4'))
+    
+    def test_get_max_index(self): 
+        ytos_slabs = generate_slabs(structure=self.ytos, 
+        hkl=1, thicknesses=[10], vacuums=[10, 20], 
+        save_slabs=False, max_size=20)
+        
+        self.assertIsNotNone(ytos_slabs)
+        self.assertEqual(len(ytos_slabs), 7)
+        self.assertWarnsRegex(UserWarning, ('Not all combinations of hkl or ' 
+        'slab/vac thicknesses were generated because of repeat structures. ' 
+        'The repeat slabs are: 001_10_20_4'))
+        self.assertWarnsRegex(UserWarning, ('Some generated slabs exceed the ' 
+        'max size specified. Slabs that exceed the max size are: 111_10_10_3, '
+        '111_10_20_3, 101_10_10_4, 101_10_20_4, 100_10_10_0, 100_10_20_0, '
+        '001_10_10_4'))
+
+    def test_get_none(self): 
+        ytos_no_slab = generate_slabs(structure=self.ytos, hkl=(0,3,5), 
+        thicknesses=[10], vacuums=[10,20], save_slabs=False)
+
+        self.assertEqual(ytos_no_slab, [])
+        self.assertWarnsRegex(UserWarning, 'No zero dipole slabs found for '
+        'specified Miller index')
+
+    def test_no_structure(self):
+        self.assertRaises(FileNotFoundError, generate_slabs, structure='waa', 
[…]
