+++ surfaxe/generation.py
@@ -1,12 +1,11 @@
 # pymatgen
-from pymatgen.core.surface import SlabGenerator, generate_all_slabs
+from pymatgen.core.surface import SlabGenerator, generate_all_slabs, get_symmetrically_distinct_miller_indices
 from pymatgen.core import Structure
 from pymatgen.core.structure import SiteCollection
 from pymatgen.symmetry.analyzer import SpacegroupAnalyzer
 
 # misc
 import warnings
-import os
 import itertools
 import functools
 import multiprocessing
@@ -15,6 +14,266 @@ import numpy as np
 
 # surfaxe
 from surfaxe.io import slabs_to_file, _custom_formatwarning
+
+def generate_slabs(structure, hkl, thicknesses, vacuums, make_fols=False, 
+make_input_files=False, max_size=500, center_slab=True, ox_states=None, 
+save_slabs=True, is_symmetric=True, layers_to_relax = None, fmt='poscar', name='POSCAR', 
+config_dict='PBEsol_config.json', user_incar_settings=None, 
+user_kpoints_settings=None, user_potcar_settings=None, parallelise=True, **kwargs): 
+    """
+    Generates all unique slabs for a specified Miller index or up to a maximum 
+    Miller index with minimum slab and vacuum thicknesses. It includes all 
+    combinations for multiple zero dipole symmetric terminations for 
+    the same Miller index. 
+    
+    The function returns None by default and generates either: 
+
+    (i) POSCAR_hkl_slab_vac_index.vasp (default) 
+    (ii) hkl/slab_vac_index folders with structure files
+    (iii) hkl/slab_vac_index with all VASP input files 
+    
+    Or if `save_slabs=False` a list of dicts of all unique slabs is returned. 
+    
+    Args:
+        structure (`str`): Filename of structure file in any format supported by 
+            pymatgen. 
+        hkl (`tuple`, `list` or `int`): Miller index as tuple, a list of Miller 
+            indices or a maximum index up to which the search should be 
+            performed. E.g. if searching for slabs up to (2,2,2) ``hkl=2``
+        thicknesses (`list`): The minimum size of the slab in Angstroms. 
+        vacuums (`list`): The minimum size of the vacuum in Angstroms. 
+        make_fols (`bool`, optional): Makes folders for each termination 
+            and slab/vacuum thickness combinations containing structure files. 
+            
+            * ``True``: A Miller index folder is created, in which folders 
+              named slab_vac_index are created to which the relevant structure 
+              files are saved. 
+                    
+                    E.g. for a (0,0,1) slab of index 1 with a slab thickness of 
+                    20 Å and vacuum thickness of 30 Å the folder structure would 
+                    be: ``001/20_30_1/POSCAR``  
+
+            * ``False``: The indexed structure files are put in a folder named  
+              after the bulk formula. 
+              
+                    E.g. for a (0,0,1) MgO slab of index 1 with a slab thickness 
+                    of 20 Å and vacuum thickness of 30 Å the folder structure 
+                    would be: ``MgO/POSCAR_001_20_30_1``
+
+            Defaults to ``False``.    
+        make_input_files (`bool`, optional): Makes INCAR, POTCAR and 
+            KPOINTS files in each folder. If ``make_input_files`` is ``True`` 
+            but ``make_files`` or ``save_slabs`` is ``False``, files will be 
+            saved to folders regardless. This only works with VASP input files, 
+            other formats are not yet supported. Defaults to ``False``. 
+        max_size (`int`, optional): The maximum number of atoms in the slab 
+            specified to raise warning about slab size. Even if the warning is 
+            raised, it still outputs the slabs regardless. Defaults to ``500``. 
[…]
