### 2021-05-14 2691e9a
add generate slabs cli and more tests
