### 2021-05-14 0d72a8d
remove get-all, get-hkl; rework load_config_dict
