+++ tests/tests_generation/test_generation.py
@@ -1,115 +1,12 @@
 import unittest
-import warnings
 import os
 import shutil
 from pathlib import Path
-from pymatgen.core import Structure
 from pymatgen.core.surface import Slab
-from surfaxe.generation import get_slabs_max_index, get_slabs_single_hkl, generate_slabs
+from surfaxe.generation import generate_slabs
 
 ytos = str(Path(__file__).parents[2].joinpath('example_data/generation/CONTCAR_conventional'))
 cdte  = str(Path(__file__).parents[2].joinpath('example_data/generation/CdTe.vasp'))
-
-class GetAllTestCase(unittest.TestCase): 
-
-    def setUp(self): 
-        self.ytos = ytos
-        self.cdte = cdte
-    
-    def test_get_max_index(self): 
-        ytos_slabs = get_slabs_max_index(structure=self.ytos, 
-        max_index=1, thicknesses=[10], vacuums=[10, 20], 
-        save_slabs=False, max_size=20)
-        
-        self.assertIsNotNone(ytos_slabs)
-        self.assertEqual(len(ytos_slabs), 7)
-        self.assertWarnsRegex(UserWarning, ('Not all combinations of hkl or ' 
-        'slab/vac thicknesses were generated because of repeat structures. ' 
-        'The repeat slabs are: 001_10_20_15'))
-        self.assertWarnsRegex(UserWarning, ('Some generated slabs exceed the ' 
-        'max size specified. Slabs that exceed the max size are: 111_10_10_3, '
-        '101_10_10_9, 100_10_10_10, 001_10_10_15, 111_10_20_3, 101_10_20_9, '
-        '100_10_20_10'))
-
-    def test_non_centrosymmetric(self): 
-        sym_true = get_slabs_max_index(structure=self.cdte, max_index=1, 
-        thicknesses=[10], vacuums=[10,20], save_slabs=False)
-        self.assertWarnsRegex(UserWarning, 'Inversion symmetry was not found '
-        'in the bulk structure, slabs produced will be non-centrosymmetric')
-
-        sym_false = get_slabs_max_index(structure=self.cdte, max_index=1, 
-        thicknesses=[10], vacuums=[10,20], save_slabs=False, is_symmetric=False)
-        self.assertEqual(len(sym_false), 2) 
-        self.assertEqual(len(sym_true), 2) 
-    
-    def test_no_structure(self): 
-        self.assertRaises(FileNotFoundError, get_slabs_max_index, structure='waa', 
-        max_index=1, thicknesses=[10], vacuums=[10,20], save_slabs=False)
-
-    def test_save_to_file(self): 
-        ytos_slabs = ytos_slabs = get_slabs_max_index(structure=self.ytos, 
-        max_index=1, thicknesses=[10], vacuums=[10, 20])
-        
-        self.assertIsNone(ytos_slabs)
-         
-
-class GetOneTestCase(unittest.TestCase): 
-
-    def setUp(self): 
-        self.ytos = ytos 
-        self.cdte = cdte
-    
-    def test_get_single_hkl(self): 
-        ytos_slab = get_slabs_single_hkl(structure=self.ytos, hkl=(0,0,1), 
-        thicknesses=[10,20], vacuums=[10,20], save_slabs=False, max_size=20)
-
-        self.assertEqual(len(ytos_slab), 1)
-        self.assertEqual(ytos_slab[0]['slab_index'], 4)
-        self.assertWarnsRegex(UserWarning, ('Some generated slabs exceed the '
-        'max size specified. Slabs that exceed the max size are: 001_10_10_4'))
[…]
