+++ setup.py
@@ -36,8 +36,6 @@ if __name__ == "__main__":
       ], 
       entry_points={
         'console_scripts': [
-          'surfaxe-getall = surfaxe.cli.getall:main', 
-          'surfaxe-gethkl = surfaxe.cli.gethkl:main', 
           'surfaxe-parsefols = surfaxe.cli.parsefols:main', 
           'surfaxe-bonds = surfaxe.cli.bonds:main', 
           'surfaxe-simplenn = surfaxe.cli.simplenn:main', 
@@ -50,7 +48,7 @@ if __name__ == "__main__":
           'surfaxe-plot-enatom = surfaxe.cli.plotenatom:main', 
           'surfaxe-plot-bonds = surfaxe.cli.plotbonds:main', 
           'surfaxe-plot-potential = surfaxe.cli.plotpotential:main', 
-          'surfaxe-generate-slabs = surfaxe.cli.generateslabs:main'
+          'surfaxe-generate = surfaxe.cli.generateslabs:main'
         ]
       }
     )

[…]
