### 2021-05-16 SMTG-UCL/surfaxe#10 @pzarabadip
[SUGGESTION] Adding pre-commit hooks for linting/formatting the code
Hi,
Thanks for developing this nice package.

While I was reading/reviwing the code, I found couple of unused imports/variables (listed below). I suggest adding `pre-commit` hooks with `pylint`, `yapf`, and `prospector` to format/lint code at commits. This would capture these and also improves readability of the code.

**Unused imports**:
* `os`,`mpl`,`plt`, `Element`, `Lattice`, and `SiteCollection` in `analysis.py`.
* `Structure` and `numpy` in `convergence.py`.
* `os` in `generation.py`.
* `mpl` in `io.py`.
* `Slab`, `Element`, `PeriodicSite` in `vasp_data.py`.
* `warnings` and `Structure` in `test_analysis.py`.
* `warnings`, `os`, `Structure`, and `Slab` in `test_convergence.py`
* `warnings`, `Structure` in `test_generation.py`
* `warnings`, `os`, `Structure`, `surfaxe.io` in `test_io.py`

**Unused variables**:
* `files` [here](https://github.com/SMTG-UCL/surfaxe/blob/b3f5fe8a40b96d326e55dc51241b93c81ec75291/surfaxe/convergence.py#L90)
* `low` and `high` [here](https://github.com/SMTG-UCL/surfaxe/blob/b3f5fe8a40b96d326e55dc51241b93c81ec75291/surfaxe/vasp_data.py#L302)

Best regards,
Pezhman
