### 2021-05-16 #3171 @pzarabadip
Hi @danielskatz.
Sorry for late response. I just finished my review on the code.

I would like to thank authors for developing `surfaxe` package which is nice and useful contribution for the community.
I carefully went through code, tutorials, tests, documentation, and paper and found from my point view that the code/paper is (almost) ready to be accepted for publication in `JOSS`. I only have few suggestions that could be nice to be implemented before publication.

* Applying formatting/linting tools on the code for which I opened an [issue](https://github.com/SMTG-UCL/surfaxe/issues/10#issue-892644619) and explained it in more detail.
* Another suggestion would be moving the CI to `GitHub Actions` to improve the sustainability in future.
* From my personal experience: it would be beneficial for the performance to use `OUTCAR` parser of [`parsevasp` package](https://github.com/aiida-vasp/parsevasp) instead of the one provided by `pymatgen`.

Best regards,
Pezhman
