### 2021-05-17 cf6155c
remove safe_load, change tests
