### 2021-05-17 #3171 @danielskatz
Thanks @pzarabadip - Do I understand correctly that these are *suggestions*, not changes that are *required* for you?

Or is the last one a requirement, tied to the fact that you didn't check the performance box in your review?
