### 2021-05-19 7d0a0d8
Add command to install surfaxe itself to tests.yml
