+++ .github/workflows/tests.yml
@@ -42,7 +42,7 @@ jobs:
       run: |
         python -m pip install --upgrade pip
         pip install pytest
-        pip install -r requirements.txt
+        pip install -r requirements.txt .
     
     - name: Test with pytest
       run: |
