+++ .travis.yml
@@ -1,15 +1,0 @@
-language: python
-
-python:
-  - "3.7"
-
-install:
-  - pip install --upgrade pip
-  - pip install -r requirements.txt .
-  
-branches:
-  only:
-  - /.*/
-
-script: 
-  - pytest
