+++ README.md
@@ -1,5 +1,5 @@
 
-![Build status](https://github.com/smtg-ucl/surfaxe/actions/workflows/tests.yml/badge.svg)
+[![Build status](https://github.com/smtg-ucl/surfaxe/actions/workflows/tests.yml/badge.svg)](https://github.com/SMTG-UCL/surfaxe/actions)
 [![Documentation Status](https://readthedocs.org/projects/surfaxe/badge/?version=latest)](https://surfaxe.readthedocs.io/en/latest/?badge=stable) 
  
 <img src='example_data/figures/surfaxe_header_v2.png' alt='surfaxe logo header' width='600'/>
