### 2021-05-19 SMTG-UCL/surfaxe#10 @brlec
Hi,

Thank you for reviewing the code and flagging these! I have now removed the unused imports in the [latest commit](https://github.com/SMTG-UCL/surfaxe/commit/d21c74a76c561c08bf5297c51b7ba3d66d7c1b75) on `develop` - these will be merged with `master` once the JOSS review is complete. Regarding unused variables, `pylint` doesn't pick up on where  `low` and `high` are referenced later on in dataframe query, so they were left in.

Best,
Kat
