### 2021-05-19 #3171 @brlec
Hi @pzarabadip,

Thank you for the thorough review!

To answer the suggestions:
- **Applying formatting/linting tools**: We have removed the unused imports in the [latest commit](https://github.com/SMTG-UCL/surfaxe/commit/d21c74a76c561c08bf5297c51b7ba3d66d7c1b75) on the `develop` branch and they will be merged with `master` once JOSS review is complete. We have added linting tools that should catch unused imports going forward.
- **Moving the CI to GitHub Actions**: We have implemented this in the [latest commit](https://github.com/SMTG-UCL/surfaxe/commit/2f39adc9f6e3d0b69dfa46d48aae84e239844243) on the `master` branch, the tests are passing.
- **Using `parsevasp`**: OUTCARs are parsed for two different purposes in `surfaxe` - getting timing information from convergence calculations or getting the core level energies. We have tested `parsevasp` vs `pymatgen` and found that `parsevasp` offers a 10x speedup for reading the timing information.
```python
from parsevasp import outcar
from pymatgen.io.vasp.outputs import Outcar
import time

# pymatgen
start = time.time()
otc = Outcar('surfaxe/example_data/convergence/Y2Ti2S2O5/001/30_50_15/OUTCAR')
times = otc.run_stats
print(time.time() - start)
>> 0.21026110649108887

# parsevasp
start = time.time()
otc = outcar.Outcar(file_path='surfaxe/example_data/convergence/Y2Ti2S2O5/001/30_50_15/OUTCAR')
times = otc.get_run_stats()
print(time.time() -  start)
>> 0.029942989349365234
```
However, the additional wait time for the user arising from `pymatgen` parser would be minimal as parsing of vasprun.xml takes far longer. `parsevasp` OUTCAR parser also cannot read in the core level energies (yet), so we are going to stick with `pymatgen` as the package for reading in OUTCARs. We will definitely look into using `parsevasp` in the future if parsing of core level energies is added.

Thanks again for the really helpful review!
Katarina
