### 2021-05-19 #3171 @danielskatz
Thanks @eihernan - re the installation problem, if it does continue, you might open an issue in https://github.com/SMTG-UCL/surfaxe and mention `https://github.com/openjournals/joss-reviews/issues/3171` there so we can track it.

And I'll look forward to a response from the authors to your other suggestions.
