+++ surfaxe/cli/cartdisp.py
@@ -1,8 +1,6 @@
 # Misc 
 from argparse import ArgumentParser
-import yaml
-import os
-import warnings 
+from ruamel.yaml import YAML
 
 # Surfaxe 
 from surfaxe.analysis import cart_displacements
@@ -37,7 +35,8 @@ def main():
 
     if args.yaml is not None: 
         with open(args.yaml, 'r') as y: 
-            yaml_args = yaml.safe_load(y)
+            yaml = YAML(typ='safe', pure=True)
+            yaml_args = yaml.load(y)
 
         cd = cart_displacements(**yaml_args)
         if ('save_csv', False) in yaml_args.items(): 
