+++ surfaxe/cli/parsefols.py
@@ -1,8 +1,8 @@
 # Misc 
 from argparse import ArgumentParser
-import yaml
 import os
-import warnings 
+
+from ruamel.yaml.main import YAML
 
 # Surfaxe 
 from surfaxe.convergence import parse_fols
@@ -55,7 +55,8 @@ def main():
 
     if args.yaml is not None: 
         with open(args.yaml, 'r') as y: 
-            yaml_args = yaml.safe_load(y)
+            yaml = YAML(typ='safe', pure=True)
+            yaml_args = yaml.load(y)
         
         parse_fols(**yaml_args)
 
