+++ surfaxe/cli/simplenn.py
@@ -1,8 +1,6 @@
 # Misc 
 from argparse import ArgumentParser
-import yaml
-import os
-import warnings 
+from ruamel.yaml.main import YAML
 from pymatgen.analysis.local_env import CrystalNN
 
 # Surfaxe 
@@ -52,7 +50,8 @@ def main():
 
     if args.yaml is not None: 
         with open(args.yaml, 'r') as y:
-            yaml_args = yaml.safe_load(y)
+            yaml = YAML(typ='safe', pure=True)
+            yaml_args = yaml.load(y)
         
         nn = simple_nn(**yaml_args)
         if ('save_csv', False) in yaml_args.items(): 
