+++ surfaxe/cli/getall.py
@@ -1,8 +1,6 @@
 # Misc 
 from argparse import ArgumentParser
-import yaml
-import os
-import warnings 
+from ruamel.yaml.main import YAML
 
 # Surfaxe 
 from surfaxe.generation import get_slabs_max_index
@@ -78,6 +76,7 @@ def main():
 
     if args.yaml is not None: 
         with open(args.yaml, 'r') as y: 
+            yaml = YAML(typ='safe', pure=True)
             yaml_args = yaml.load(y)
 
         get_slabs_max_index(**yaml_args)
