+++ surfaxe/cli/potential.py
@@ -1,8 +1,7 @@
 # Misc 
 from argparse import ArgumentParser
-import yaml
-import os
-import warnings 
+
+from ruamel.yaml.main import YAML
 
 # Surfaxe 
 from surfaxe.analysis import electrostatic_potential 
@@ -46,7 +45,8 @@ def main():
 
     if args.yaml is not None: 
         with open(args.yaml, 'r') as y:
-            yaml_args = yaml.safe_load(y)
+            yaml = YAML(typ='safe', pure=True)
+            yaml_args = yaml.load(y)
 
         ep = electrostatic_potential(**yaml_args)
         if ('save_csv', False) in yaml_args.items(): 
