### 2021-05-19 6dbdcb6
more info on pymatgen potcars
