### 2021-05-20 649ae3d
change to 3.7, add better soft check for potcars
