+++ setup.py
@@ -27,6 +27,7 @@ if __name__ == "__main__":
       packages=['surfaxe'],
       zip_safe=False,
       install_requires=['scipy', 'numpy', 'spglib', 'pymatgen','pandas'],
+      python_requires='>=3.7',
       classifiers=[
         'Programming Language :: Python',
         'Development Status :: Development',
