### 2021-05-20 #3171 @brlec
Hi @eihernan,

Thank you for your review and kind words.

Thank you for raising the issue with CLI - we believe this was related to how yaml was imported in the CLI scripts. We have replaced `yaml` with `ruamel.yaml` which should work now even in blank virtual environments as `ruamel.yaml` is a dependency of `pymatgen`.

To respond to suggestions:
- **POTCAR set up for `pymatgen`**: `Pymatgen` is fairly rigid with respect to how it can read input files, unfortunately this means the POTCARs need to be additionally set up to work with `pymatgen`. In response to the confusion this may cause we have added the following line to README, with a link to the `pymatgen` documentation that outlines how to complete the set up.

"_For the code to generate VASP input files along with the surface slabs, POTCARs need to be [set up with pymatgen](https://pymatgen.org/installation.html#potcar-setup)_"

We also added a similar line to the Installation section of surfaxe docs.

Additionally, we modified the code to check that `PMG_VASP_PSP_DIR` key exists in `pymatgen.core.SETTINGS`
before structures are written to file. If POTCARs are not set up, only the structures are written to file and the following warning is raised: `POTCAR directory not set up in pymatgen, only POSCARs were generated`. We are open to suggestions on how this could be further improved.

- **Configuration dictionaries**: We agree that for someone not familiar with `pymatgen` ecosystem this may be somewhat confusing. We added a section in the Tutorials section of documentation on the [`develop `branch](https://github.com/SMTG-UCL/surfaxe/tree/develop) outlining how this should be used.

All changes were made on the [`develop` branch](https://github.com/SMTG-UCL/surfaxe/tree/develop) if you would like to try re-installing the package in a virtual environment to see if the CLI problem is fixed now?

Thank you again,
Katarina
