### 2021-05-20 #3171 @brlec
Hi @danielskatz - the JOSS release is tagged [v1.2](https://github.com/SMTG-UCL/surfaxe/tree/v1.2) and was archived on Zenodo with the correct metadata. The DOI is 10.5281/zenodo.4776449
