### 2021-05-20 #3171 @danielskatz
@brlec - I'm suggesting some small changes to the paper in https://github.com/SMTG-UCL/surfaxe/pull/11

Please either merge this or let me know what you disagree with.
