### 2021-05-20 #3171 @brlec
@danielskatz we've merged all the suggestions
